Update readme to send folks to the wiki instead
The readme is really just for meta information about the repo, and shouldn't have anything to do with the repo contents at this point. Therefore, this meta information would be better stored in the wiki where there is only one relevant branch, instead of the code base which could have multiple active branches at any point.