Revert some doc changes from commit 2102cca10848
As per @mradamcox's and @ekansa's review in PR #345, the changes to
the workflow documentation in commit 2102cca10848 were incorrect; this
reverts those changes.  The fixes to the import/export section still
look right, however -- I checked them over again in light of the
review -- so this commit leaves them in place.